requirements.txt
# ============================================
# Agno Agent Requirements
# ============================================
# Core dependencies for Agno AI agent with OpenTelemetry tracing
# Generated from: conda env agno-env

# === Core Agent Framework ===
agno==2.4.7

# === OpenAI LLM ===
openai>=1.0.0

# === OpenTelemetry Tracing ===
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-semantic-conventions==0.60b1

# === Agno Instrumentation ===
openinference-instrumentation-agno==0.1.27

# === Environment & Configuration ===
python-dotenv==1.2.1

# === Streamlit Web Interface ===
streamlit>=1.30.0

# ============================================
# Optional Dependencies
# ============================================

# === Agent Lightning (for training) ===
agentlightning>=0.3.0

# === OTLP Instrumentation ===
openinference-instrumentation-openai>=0.1.0

# === Development Tools ===
# Uncomment for development
# pytest>=7.0.0
# black>=23.0.0
# ruff>=0.1.0

